Enchantes on single shard replication database names
Now when Stampede is replicating from a single shard, it uses the non
sharding mode mode, which means that database names, loggers, metrics,
etc will be the same as when we didn't support sharding.
